+++ src/service/nepmeds-doctor-availability.ts
@@ -40,9 +40,58 @@ export const useCreateDoctorAvailability = () => {
   >(createDoctorAvailability, {
     onSuccess: () => {
       queryClient.invalidateQueries(api.doctor_availability);
-      queryClient.fetchQuery(api.doctor_availability);
     },
   });
 
   return mutation;
+};
+
+export const getSingleAvailability = async (id: number) => {
+  const response = await HttpClient.get<
+    NepMedsResponse<IGetDoctorAvailability>
+  >(api.doctor_availability + `${id}/`);
+
+  return response.data.data;
+};
+
+export const updateDoctorAvailability = async ({
+  id,
+  data,
+}: {
+  id: number;
+  data: IGetDoctorAvailability;
+}) => {
+  const response = await HttpClient.patch<NepMedsResponse>(
+    api.doctor_availability + id + "/",
+    data
+  );
+
+  return response;
+};
+
+export const useUpdateDoctorAvailability = () => {
+  const queryClient = useQueryClient();
+
+  return useMutation(api.doctor_availability, updateDoctorAvailability, {
+    onSuccess: () => {
+      queryClient.invalidateQueries(api.doctor_availability);
+    },
+  });
+};
+
+const deleteAvailability = async (availabilityId: { id: number }) => {
+  const response = await HttpClient.delete<NepMedsResponse>(
+    api.doctor_availability + availabilityId.id + "/"
+  );
+  return response;
+};
+
+export const useDeleteAvailability = () => {
+  const queryClient = useQueryClient();
+
+  return useMutation(deleteAvailability, {
+    onSuccess: () => {
+      queryClient.invalidateQueries(api.doctor_availability);
+    },
+  });
 };